VIH-6436 new get non-closed conference by hearing id update GetConferenceByHearingRefIdQuery.cs to ignore status
